🐛 Correctly exhaust lazy sequence.
Errors are now surfacing correctly from the parser.

I know that I am going with this mutable module style of code here to make it easier to follow
along with the book. Just noting now that there are already problems with test concurrency
since the parser is basically a shared singleton object.

Shouldn't be a problem in the short term since there will only every be one parser instance.

This style is has the best parts of OO in my opinion, limited and controlled mutation through
atoms, no private data.